rl: fixes from high-effort code review of the RL stack
Confirmed + fixed:
- bc_data: is_comet was NEVER set for mid-game comets (read from the
  always-empty step-0 comet_planet_ids); comets trained as rotating
  planets through every comet phase. Now flagged in the wiring loop.
- ppo advantage norm: floor the denominator at 1e-3 (not +1e-8) so
  near-constant league chunks can't inflate noise-level advantages ~1e4.
- numpy_infer launch ranking: np.argsort(kind='stable') to match
  jnp.argsort, so eval keeps the same fleets as training under ties.
- numpy_infer num_agents: made MONOTONIC (re-evaluate each turn, never
  downgrade) so a mid-game first-obs / early elimination can only raise
  the seat count.
- export_agent: fp16 -> fp32 params. Review measured fp16 flipping the
  argmax target on ~0.8% of turns, invisible to the fp64 parity tests.
  File 665KB, decision-identical, plays clean.
- new test_obs_to_arrays_comet_positions_match_engine: guards the comet
  branch of obs_to_arrays that the GameState parity tests bypass
  (seed-robust). 7/7 numpy-parity + 5/5 aim tests green.

Refuted by verification (no change):
- comet path/planet_ids desync: the 4 comets in a group share a path
  length (4-fold symmetry) so they expire together; planet_ids/paths
  never desync in practice. Kept a harmless alive-paths guard anyway.
- num_agents 4P-blindness: first act() is step 0 where all home planets
  reveal the true count; monotonic guard added for defense-in-depth.

+++ rl/export_agent.py
@@ -53,8 +53,13 @@ def export(ckpt_path: str, out_path: str):
         d = pickle.load(f)
     params = {k: np.asarray(v) for k, v in d["params"].items()}
 
+    # fp32, not fp16: a prior review measured fp16 round-trip flipping
+    # the argmax target on ~0.8% of turns (borderline logit gaps) — the
+    # fp64 parity tests never see it. fp32 keeps the exported agent
+    # decision-identical to the trained policy; zlib still gets the file
+    # to ~0.6 MB, far under the Kaggle limit.
     buf = io.BytesIO()
-    np.savez(buf, **{k: v.astype(np.float16) for k, v in params.items()})
+    np.savez(buf, **{k: v.astype(np.float32) for k, v in params.items()})
     blob = zlib.compress(buf.getvalue(), 9)
     b64 = base64.b64encode(blob).decode()
     lines = [f'    "{b64[i:i + 96]}"' for i in range(0, len(b64), 96)]
